bugfix: DB apply_block must run concurrently
`deadpool`'s method `interact` is blocking, it waits for the closure to
complete its execution before returning. This causes a deadlock, because
of the way state synchronization is performed in the store server.

In the store, the transaction is opened and data is written to the WAL
file, but not immedially committed. Before comiting the data, the
in-memory datastructures prevent read operations. This meanst the server
has to acquire the exclusive write lock. This operation demands
concurrency and synchronization among the DB and state components. This
required a change from a blocking to a concurrent API.

`deadpool` doesn't seem to have a native concurrent API, so a async task
was used instead. `tokio` does not support scoped tasks, so thread local
state can not be used across tasks, for this reason all the data was
modified from being borrowed to being owned, so that the data can be
moved in to the async task.