Connect 4 nodes in block-producer, and create `miden-node` binary (#85)
* fix tx_queue lifetime issue

* gitignore .vscode

* remove allow(dead_code)

* db -> store

* DefaultStore

* BlockProducerConfig::store_endpoint

* initialize DefaultStore with grpc store client

* add conversion domain -> raw for BlockHeader

* impl apply_block: block_header

* apply_block: accounts

* apply_block: all but notes

* get_tx_inputs

* get_block_inputs

* add TODO

* miden-node crate

* logging

* miden-node: start store

* start rpc

* start block producer

* block_producer: Add proven transaction payload (#89)

* implement apply_block

* store: errors are part of the lib, not the binary

* add tracing dep to node

* fix block producer config

* start block producer first

* txqueue: create batch when empty queue

* also start batch builder in server

* (empty) block building works

* genesis_header function

* store apply_block: return genesis header if store empty

* genesis block header: use timestamp = 0

* nullifier root: use empty TieredSmt

* clippy

* added component tag, to differentiate serve initialization log lines

* tracing: enable json format and inherit features on workspace

* logging: add env_filter

It allows to use the env variable `RUST_LOG=<level>` to set the logging
level.

* logging: enable logging of spans

* Revert "txqueue: create batch when empty queue"

This reverts commit 38c19ceff0c03925aae8776c63f99b814f7d26c4.

* bugfix: DB apply_block must run concurrently

`deadpool`'s method `interact` is blocking, it waits for the closure to
complete its execution before returning. This causes a deadlock, because
of the way state synchronization is performed in the store server.

In the store, the transaction is opened and data is written to the WAL
file, but not immedially committed. Before comiting the data, the
in-memory datastructures prevent read operations. This meanst the server
has to acquire the exclusive write lock. This operation demands
concurrency and synchronization among the DB and state components. This
required a change from a blocking to a concurrent API.

`deadpool` doesn't seem to have a native concurrent API, so a async task
was used instead. `tokio` does not support scoped tasks, so thread local
state can not be used across tasks, for this reason all the data was
modified from being borrowed to being owned, so that the data can be
moved in to the async task.

* clippy

* fix doc link

* apply_block accounts: use convert

* nullifiers: use convert

* convert (u64, NoteEnvelope) to NoteCreated

* remove `constants` module

* block producer: make some modules private

* remove `constants` module

* remove constants mod

* move batch_builder constants to lib.rs

* miden-node: update package in Cargo.toml

* miden-node: update package in Cargo.toml

* DefaultStateView

* define constants for block producer server

* remove pub(crate)

---------

Co-authored-by: Augusto Hack <hack.augusto@gmail.com>