feat: component startup is now coherent (#488)
Splits node component startup into two steps: initialization and serving the api.

Initialization includes binding the tcp listener on the endpoint. This enables the caller to know when the component is
ready for communication prior to actually spawning the server.

In particular, this solves the issue of spawning a server, and then sleeping an arbitrary amount to wait for the server to 
become available. Instead we now do `Component::init` after which the endpoint is valid.